implementing...

+++ gobot_base/src/odom.cpp
@@ -70,7 +70,7 @@ int main(int argc, char** argv){
                 double right_vel = right_dist / dt;
 
                 //compute odometry in a typical way given the velocities of the robot
-                double vel = (right_vel + left_vel) / (double) 2;
+                double vel = (right_vel + left_vel) / 2.0;
                 double vx = vel * cos(th);
                 double vy = vel * sin(th);
                 double vth = (right_vel - left_vel) / wheel_separation;
